# openapi.yaml
openapi: 3.0.3
info:
  title: Nutri Thata API
  description: API para plataforma de educação nutricional com sistema de módulos, convites e upload de mídia
  version: 1.0.0
  contact:
    name: API Support
    email: support@nutrithata.com
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://api.nutrithata.com
    description: Produção
security:
  - BearerAuth: []
paths:
...
  /users/me:
    get:
      tags:
        - Usuários
      summary: Obter perfil do usuário
      description: Dados pessoais do usuário autenticado
      responses:
        '200':
          description: Perfil do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Usuários
      summary: Atualizar perfil do usuário
      description: Atualiza dados pessoais do usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "João Silva Santos"
                phone:
                  type: string
                  example: "(11) 98888-8888"
                birthDate:
                  type: string
                  format: date
                  example: "1990-01-01"
      responses:
        '200':
          description: Perfil atualizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Perfil atualizado com sucesso"
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== PROFISSIONAIS ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        faultstring:
          type: string
          example: "Credenciais inválidas"
        stacktrace:
          type: string
          description: "Disponível apenas em ambiente de desenvolvimento"
          example: "com.nutrithata.exception.AuthenticationException: Invalid credentials..."

    # ========== MODELOS DE USUÁRIO ==========
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao@email.com"
        role:
          type: string
          enum: ["PATIENT", "PROFESSIONAL", "ADMIN"]
          example: "PATIENT"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            phone:
              type: string
              example: "(11) 99999-9999"
            birthDate:
              type: string
              format: date
              example: "1990-01-01"
...